src/detection/sigma_engine.py
import os
import yaml
from typing import List, Dict, Any
# pySigma imports - commented out for MVP due to upstream dependency issues
# from sigma.collection import SigmaCollection
# from sigma.rule import SigmaRule
# from sigma.evaluator import SigmaEvaluator

class SimpleEvaluator:
    """
    A simplified evaluator since pySigma's default StructureEvaluator 
    might need specific backend integration. We can roll a simple one 
    checking key-value pairs for the MVP.
    """
    def __init__(self, rules: List[Any]):
        self.rules = rules

    def match(self, log: Dict[str, Any]) -> List[Any]:
        matches = []
        # for rule in self.rules:
        #    if self._apply_detection(rule.detection, log):
        #        matches.append(rule)
        return matches

    def _apply_detection(self, detection: Any, log: Dict[str, Any]) -> bool:
        return False

class SigmaEngine:
    def __init__(self, config: Dict[str, Any]):
        self.rules_path = config.get('rules_path', '')
        # self.collection = self._load_rules()
        self.backend = None 

    def _load_rules(self) -> Any:
        # Placeholder
        if not os.path.exists(self.rules_path):
            print(f"Warning: Rules path {self.rules_path} does not exist.")
            return []
        return []

    def evaluate(self, logs: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
        """
        Evaluates the batch of logs against loaded rules.
        Returns a list of generated alerts.
        """
        alerts = []
        
        # Simple manual check demo
        for log in logs:
            # Manual hardcoded rule for MVP verification until full engine is connected
            if str(log.get('EventID')) == '4625':
                alerts.append({
                    'rule_title': 'Failed Login Attempt',
                    'rule_level': 'high',
                    'log_data': log
                })
                
        return alerts
