requirements.txt
clickhouse-driver==0.2.6
requests==2.31.0
PyYAML==6.0.1
pydantic==2.5.2
pydantic
pydantic

